// src/api-client/models/location-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * What kind of site this is.
 * @export
 * @enum {string}
 */

export const LocationType = {
    Store: 'STORE',
    Warehouse: 'WAREHOUSE',
    DistributionCenter: 'DISTRIBUTION_CENTER',
    Kitchen: 'KITCHEN',
    Office: 'OFFICE',
    PickupPoint: 'PICKUP_POINT'
} as const;

export type LocationType = typeof LocationType[keyof typeof LocationType];